eia2tidy

Extracts data from the Energy Information Administration (EIA) API to tibble format with optional custom series name. Makes a clean wrapper for use with purrr for multiple series extraction. Query Browser at https://www.eia.gov/opendata/qb.php.

eia2tidy(ticker, key, name = " ")

Arguments

ticker
EIA series name.

key
Your private EIA API token as character "<yourapikey>".

name
Name you want to give the series. Defaults to ticker if set to " "

Value

A tibble object with class date for weekly, monthly, quarterly or annual data and class POSIXct for hourly.

Author

Philippe Cote

Examples

if (FALSE) {
# Single Series
RTL::eia2tidy(ticker = "PET.MCRFPTX2.M", key = "yourapikey", name = "TexasProd")
# Multiple Series
eia_df <-tibble::tribble(~ticker, ~name,
"PET.W_EPC0_SAX_YCUOK_MBBL.W", "CrudeCushing",
"NG.NW2_EPG0_SWO_R48_BCF.W","NGLower48") %>%
 dplyr::mutate(key = "EIAkey") %>%
 dplyr::mutate(df = purrr::pmap(list(ticker,key,name),.f=RTL::eia2tidy)) %>%
 dplyr::select(df) %>% tidyr::unnest(df)
 }