Computes betas of futures contracts with respect to the 1st line contract

Returns betas of futures contracts versus front futures contract.

promptBeta(x = x, period = "all", betatype = "all", output = "chart")

Arguments

x

Wide dataframe with date column and multiple series columns (multivariate).

period

"all" or numeric period of time in last n periods as character eg "100".

betatype

"all" "bull" "bear".

output

"betas" or "chart"

Value

betas data frame or plotly chart of betas

Author

Philippe Cote

Examples

x <- dflong %>%
 dplyr::filter(grepl("CL",series)) %>%
 dplyr::mutate(series = readr::parse_number(series)) %>% dplyr::group_by(series) %>%
 RTL::returns(df = ., retType = "abs",period.return = 1,spread = TRUE) %>%
 RTL::rolladjust(x = .,commodityname = c("cmewti"),rolltype = c("Last.Trade")) %>%
 # removing the day it prices went negative...
 dplyr::filter(!date %in% c(as.Date("2020-04-20"),as.Date("2020-04-21")))
promptBeta(x = x, period = "all", betatype = "all", output = "chart")

promptBeta(x = x, period = "all", betatype = "all", output = "betas")
#>          all      bull      bear contract
#> 1  1.0000000 1.0000000 1.0000000        1
#> 2  0.9508543 0.9025136 0.9722592        2
#> 3  0.9169546 0.8622822 0.9457595        3
#> 4  0.8857040 0.8251581 0.9194247        4
#> 5  0.8575744 0.7926277 0.8949486        5
#> 6  0.8325908 0.7651637 0.8722918        6
#> 7  0.8100888 0.7411899 0.8520433        7
#> 8  0.7897918 0.7207749 0.8335220        8
#> 9  0.7706979 0.7020625 0.8161397        9
#> 10 0.7530055 0.6852524 0.7997696       10
#> 11 0.7361606 0.6693021 0.7842289       11
#> 12 0.7206226 0.6545388 0.7698706       12
#> 13 0.7061510 0.6408662 0.7566623       13
#> 14 0.6925537 0.6283004 0.7439051       14
#> 15 0.6794230 0.6161885 0.7315564       15
#> 16 0.6666696 0.6043700 0.7191363       16
#> 17 0.6541838 0.5931041 0.7066943       17
#> 18 0.6420850 0.5820031 0.6944254       18
#> 19 0.6304195 0.5714521 0.6826029       19
#> 20 0.6190612 0.5615031 0.6707071       20
#> 21 0.6082039 0.5519779 0.6593465       21
#> 22 0.5977135 0.5428863 0.6482350       22
#> 23 0.5875812 0.5339498 0.6373788       23
#> 24 0.5779219 0.5252471 0.6270625       24
#> 25 0.5684987 0.5170047 0.6170211       25
#> 26 0.5593533 0.5086210 0.6075189       26
#> 27 0.5507602 0.5009623 0.5986735       27
#> 28 0.5425671 0.4934726 0.5903698       28
#> 29 0.5348067 0.4859396 0.5819580       29
#> 30 0.5263497 0.4787077 0.5762733       30
#> 31 0.5185630 0.4734996 0.5687889       31
#> 32 0.5113867 0.4676721 0.5615521       32
#> 33 0.5052875 0.4609526 0.5543029       33
#> 34 0.4990401 0.4564023 0.5478123       34
#> 35 0.4936763 0.4517869 0.5412591       35
#> 36 0.4874319 0.4475811 0.5365569       36
promptBeta(x = x, period = "100", betatype = "all", output = "betas")
#>          all      bull      bear contract
#> 1  1.0000000 1.0000000 1.0000000        1
#> 2  0.9593291 0.9453227 0.9557342        2
#> 3  0.9174201 0.8739853 0.9210583        3
#> 4  0.8792506 0.8126870 0.8921204        4
#> 5  0.8438497 0.7564847 0.8652213        5
#> 6  0.8108745 0.7113232 0.8382109        6
#> 7  0.7816495 0.6756198 0.8127370        7
#> 8  0.7556121 0.6498369 0.7881054        8
#> 9  0.7314733 0.6267341 0.7638175        9
#> 10 0.7106253 0.6095867 0.7408056       10
#> 11 0.6901812 0.5909606 0.7210135       11
#> 12 0.6709214 0.5779832 0.6994330       12
#> 13 0.6529081 0.5650347 0.6801496       13
#> 14 0.6356092 0.5516906 0.6617423       14
#> 15 0.6185284 0.5393102 0.6429343       15
#> 16 0.6014011 0.5273853 0.6221137       16
#> 17 0.5846395 0.5149339 0.6031110       17
#> 18 0.5689683 0.5028332 0.5854410       18
#> 19 0.5538252 0.4924031 0.5684681       19
#> 20 0.5394423 0.4819601 0.5519717       20
#> 21 0.5258178 0.4706177 0.5369548       21
#> 22 0.5126865 0.4623782 0.5222177       22
#> 23 0.5002300 0.4513120 0.5086026       23
#> 24 0.4892264 0.4411467 0.4964867       24
#> 25 0.4780418 0.4353564 0.4828892       25
#> 26 0.4663521 0.4255748 0.4706854       26
#> 27 0.4550628 0.4152572 0.4582306       27
#> 28 0.4456259 0.4070876 0.4484312       28
#> 29 0.4352863 0.4012696 0.4348532       29
#> 30 0.4252269 0.3895488 0.4245715       30
#> 31 0.4161279 0.3830514 0.4143669       31
#> 32 0.4066380 0.3773883 0.4036410       32
#> 33 0.3976167 0.3711592 0.3928561       33
#> 34 0.3891906 0.3635363 0.3841342       34
#> 35 0.3801192 0.3573875 0.3727765       35
#> 36 0.3713941 0.3502187 0.3636087       36