returns

Computes periodic returns from a dataframe ordered by date

returns(df = dflong, retType = "abs", period.return = 1, spread = FALSE)

Arguments

df
Long dataframe with colnames = c("date","value","series")

retType
"abs" for absolute, "rel" for relative, or "log" for log returns.

period.return
Number of rows over which to compute returns.

spread
TRUE if you want to spread into a long dataframe.

Value

A dataframe object of returns.

Author

Philippe Cote

Examples

returns(df = dflong,retType = "abs",period.return = 1,spread = TRUE)

#> # A tibble: 4,016 × 169
#>    date       ALI01 ALI02 ALI03 ALI04 ALI05 ALI06 AUP01 AUP02 AUP03 AUP04 AUP05
#>    <date>     <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 2006-02-27    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  2 2006-02-28    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  3 2006-03-01    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  4 2006-03-02    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  5 2006-03-03    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  6 2006-03-06    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  7 2006-03-07    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  8 2006-03-08    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  9 2006-03-09    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> 10 2006-03-10    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> # … with 4,006 more rows, and 157 more variables: AUP06 <dbl>, BRN01 <dbl>,
#> #   BRN02 <dbl>, BRN03 <dbl>, BRN04 <dbl>, BRN05 <dbl>, BRN06 <dbl>,
#> #   BRN07 <dbl>, BRN08 <dbl>, BRN09 <dbl>, BRN10 <dbl>, BRN11 <dbl>,
#> #   BRN12 <dbl>, BRN13 <dbl>, BRN14 <dbl>, BRN15 <dbl>, BRN16 <dbl>,
#> #   BRN17 <dbl>, BRN18 <dbl>, BRN19 <dbl>, BRN20 <dbl>, BRN21 <dbl>,
#> #   BRN22 <dbl>, BRN23 <dbl>, BRN24 <dbl>, BRN25 <dbl>, BRN26 <dbl>,
#> #   BRN27 <dbl>, BRN28 <dbl>, BRN29 <dbl>, BRN30 <dbl>, BRN31 <dbl>, …

returns(df = dflong,retType = "abs",period.return = 1,spread = FALSE)

#> # A tibble: 554,627 × 4
#> # Groups:   series [168]
#>    date       series value returns
#>    <date>     <chr>  <dbl>   <dbl>
#>  1 2006-02-27 BRN01   61.0   -1.61
#>  2 2006-02-27 BRN02   61.8   -1.59
#>  3 2006-02-27 BRN03   62.5   -1.57
#>  4 2006-02-27 BRN04   63.1   -1.52
#>  5 2006-02-27 BRN05   63.5   -1.51
#>  6 2006-02-27 BRN06   63.9   -1.53
#>  7 2006-02-27 BRN07   64.2   -1.52
#>  8 2006-02-27 BRN08   64.5   -1.49
#>  9 2006-02-27 BRN09   64.7   -1.46
#> 10 2006-02-27 BRN10   65.0   -1.45
#> # … with 554,617 more rows