Adjusts daily returns for futures contracts roll

Returns a xts price or return object adjusted for contract roll. The methodology used to adjust returns is to remove the daily returns on the day after expiry and for prices to adjust historical rolling front month contracts by the size of the roll at each expiry. This is conducive to quantitative trading strategies as it reflects the PL of a financial trader.

rolladjust(x, commodityname = c("cmewti"), rolltype = c("Last.Trade"), ...)

Arguments

x

A df of returns.

commodityname

Name of commodity in expiry_table: unique(expiry_table$cmdty) or "cmecan" for WCW

rolltype

Type of contract roll: "Last.Trade" or "First.Notice".

...

Other parms

Value

Roll-adjusted xts object of returns

Author

Philippe Cote

Examples

ret <- dplyr::tibble(date = seq.Date(Sys.Date() - 60, Sys.Date(), 1), CL01 = rnorm(61, 0, 1))
rolladjust(x = ret, commodityname = c("cmewti"), rolltype = c("Last.Trade"))
#> # A tibble: 58 × 2
#>    date         CL01
#>    <date>      <dbl>
#>  1 2022-03-10  1.41 
#>  2 2022-03-11  1.69 
#>  3 2022-03-12  1.51 
#>  4 2022-03-13  0.508
#>  5 2022-03-14 -0.502
#>  6 2022-03-15 -1.40 
#>  7 2022-03-16 -0.541
#>  8 2022-03-17 -1.83 
#>  9 2022-03-18  0.579
#> 10 2022-03-19  0.914
#> # … with 48 more rows