Multivariate Normal from historical dataset

Generates multivariate random epsilons using absolute returns.

simMultivariates(nsims = 10, x, s0 = NULL)

Arguments

nsims

Number of simulations. Defaults to 10

x

Wide data frame of prices with date as first column.

s0

Vector of starting value for each variables. Defaults to NULL with zero.

Value

List of means, sds, covariance matrix, correlation matrix and simulated values

Author

Philippe Cote

Examples

simMultivariates(nsims = 10, x = RTL::fizdiffs, s0 = NULL)
#> $ave
#> # A tibble: 1 × 15
#>   WTI.CMA01 WTI.CM…¹  WCS.CUS  WCS.HOU  WCS.HDY   CL.EDM MSW.EDM Mars.…² Mars.…³
#>       <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>   <dbl>
#> 1    0.0229   0.0217 -0.00327 -0.00311 -0.00408 -2.21e-5 0.00256 0.00192 0.00123
#> # … with 6 more variables: LLS.SJA <dbl>, SYN.EDM <dbl>, C5.EDM <dbl>,
#> #   DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>, and abbreviated variable names
#> #   ¹​WTI.CMA02, ²​Mars.CLO01, ³​Mars.CLO02
#> 
#> $sd
#> # A tibble: 1 × 15
#>   WTI.C…¹ WTI.C…² WCS.CUS WCS.HOU WCS.HDY CL.EDM MSW.EDM Mars.…³ Mars.…⁴ LLS.SJA
#>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>  <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1    1.63    1.57   0.462   0.482    1.06 0.0724   0.877   0.675   0.384    1.75
#> # … with 5 more variables: SYN.EDM <dbl>, C5.EDM <dbl>, DUB.EFS <dbl>,
#> #   WTI.MEH <dbl>, WTI.MID <dbl>, and abbreviated variable names ¹​WTI.CMA01,
#> #   ²​WTI.CMA02, ³​Mars.CLO01, ⁴​Mars.CLO02
#> 
#> $corMat
#>                WTI.CMA01     WTI.CMA02     WCS.CUS      WCS.HOU     WCS.HDY
#> WTI.CMA01   1.0000000000  0.9599401823 -0.01514799 -0.001257599 -0.03531737
#> WTI.CMA02   0.9599401823  1.0000000000 -0.02110985 -0.005345448 -0.03765458
#> WCS.CUS    -0.0151479890 -0.0211098532  1.00000000  0.639161893  0.20788002
#> WCS.HOU    -0.0012575991 -0.0053454485  0.63916189  1.000000000  0.18508175
#> WCS.HDY    -0.0353173730 -0.0376545810  0.20788002  0.185081749  1.00000000
#> CL.EDM     -0.0206004052 -0.0213508853 -0.01197243 -0.005783470 -0.01512863
#> MSW.EDM     0.0010513198  0.0024164584  0.28926147  0.255254031  0.30388345
#> Mars.CLO01  0.0043960397  0.0094447831  0.30578511  0.338748136  0.15413667
#> Mars.CLO02 -0.0016931393  0.0001878972  0.34714617  0.386765758  0.18872527
#> LLS.SJA     0.0003823935  0.0044284306  0.28927087  0.324050354  0.16044880
#> SYN.EDM     0.0247721984  0.0240464028  0.27226776  0.252012602  0.30873754
#> C5.EDM      0.0016101472  0.0007756348  0.25462763  0.253589805  0.25648283
#> DUB.EFS    -0.0444873993 -0.0380083146  0.27812000  0.233980570  0.19453393
#> WTI.MEH    -0.0006523393  0.0025844084  0.32031335  0.356815349  0.17557680
#> WTI.MID     0.0181882058  0.0194030033  0.27518107  0.286963293  0.16804688
#>                  CL.EDM      MSW.EDM   Mars.CLO01    Mars.CLO02      LLS.SJA
#> WTI.CMA01  -0.020600405  0.001051320  0.004396040 -0.0016931393 0.0003823935
#> WTI.CMA02  -0.021350885  0.002416458  0.009444783  0.0001878972 0.0044284306
#> WCS.CUS    -0.011972430  0.289261473  0.305785114  0.3471461724 0.2892708706
#> WCS.HOU    -0.005783470  0.255254031  0.338748136  0.3867657577 0.3240503538
#> WCS.HDY    -0.015128629  0.303883454  0.154136668  0.1887252657 0.1604488005
#> CL.EDM      1.000000000 -0.006513593 -0.009944355 -0.0221546284 0.0014132159
#> MSW.EDM    -0.006513593  1.000000000  0.263831483  0.3314302060 0.2942996266
#> Mars.CLO01 -0.009944355  0.263831483  1.000000000  0.6609371582 0.5375658964
#> Mars.CLO02 -0.022154628  0.331430206  0.660937158  1.0000000000 0.5340954224
#> LLS.SJA     0.001413216  0.294299627  0.537565896  0.5340954224 1.0000000000
#> SYN.EDM    -0.009452395  0.724190164  0.246085427  0.3088754466 0.2672071598
#> C5.EDM     -0.002403160  0.457471851  0.238783297  0.3062369433 0.2883780982
#> DUB.EFS     0.005211843  0.300929794  0.255542960  0.3219458310 0.2591543059
#> WTI.MEH     0.008170118  0.341432145  0.506919592  0.5486991288 0.5761865672
#> WTI.MID    -0.015100909  0.279799874  0.348366283  0.3788468389 0.3829266893
#>                 SYN.EDM        C5.EDM      DUB.EFS       WTI.MEH     WTI.MID
#> WTI.CMA01   0.024772198  0.0016101472 -0.044487399 -0.0006523393  0.01818821
#> WTI.CMA02   0.024046403  0.0007756348 -0.038008315  0.0025844084  0.01940300
#> WCS.CUS     0.272267760  0.2546276326  0.278119996  0.3203133483  0.27518107
#> WCS.HOU     0.252012602  0.2535898051  0.233980570  0.3568153489  0.28696329
#> WCS.HDY     0.308737541  0.2564828334  0.194533926  0.1755767983  0.16804688
#> CL.EDM     -0.009452395 -0.0024031598  0.005211843  0.0081701179 -0.01510091
#> MSW.EDM     0.724190164  0.4574718507  0.300929794  0.3414321451  0.27979987
#> Mars.CLO01  0.246085427  0.2387832974  0.255542960  0.5069195922  0.34836628
#> Mars.CLO02  0.308875447  0.3062369433  0.321945831  0.5486991288  0.37884684
#> LLS.SJA     0.267207160  0.2883780982  0.259154306  0.5761865672  0.38292669
#> SYN.EDM     1.000000000  0.4354774442  0.275064584  0.3201483403  0.26796392
#> C5.EDM      0.435477444  1.0000000000  0.298514852  0.3346339226  0.26612511
#> DUB.EFS     0.275064584  0.2985148522  1.000000000  0.2982975910  0.27419739
#> WTI.MEH     0.320148340  0.3346339226  0.298297591  1.0000000000  0.51708207
#> WTI.MID     0.267963918  0.2661251132  0.274197394  0.5170820694  1.00000000
#> 
#> $coVaR
#> # A tibble: 15 × 16
#>    series     WTI.CMA01 WTI.CMA02   WCS.CUS   WCS.HOU  WCS.HDY   CL.EDM  MSW.EDM
#>    <chr>          <dbl>     <dbl>     <dbl>     <dbl>    <dbl>    <dbl>    <dbl>
#>  1 WTI.CMA01   2.66      2.45     -0.0114   -0.000989 -0.0610  -2.43e-3  1.50e-3
#>  2 WTI.CMA02   2.45      2.45     -0.0153   -0.00403  -0.0624  -2.42e-3  3.32e-3
#>  3 WCS.CUS    -0.0114   -0.0153    0.213     0.142     0.102   -4.00e-4  1.17e-1
#>  4 WCS.HOU    -0.000989 -0.00403   0.142     0.232     0.0944  -2.02e-4  1.08e-1
#>  5 WCS.HDY    -0.0610   -0.0624    0.102     0.0944    1.12    -1.16e-3  2.82e-1
#>  6 CL.EDM     -0.00243  -0.00242  -0.000400 -0.000202 -0.00116  5.25e-3 -4.13e-4
#>  7 MSW.EDM     0.00150   0.00332   0.117     0.108     0.282   -4.13e-4  7.68e-1
#>  8 Mars.CLO01  0.00484   0.00998   0.0954    0.110     0.110   -4.86e-4  1.56e-1
#>  9 Mars.CLO02 -0.00106   0.000113  0.0615    0.0715    0.0767  -6.16e-4  1.11e-1
#> 10 LLS.SJA     0.00109   0.0121    0.234     0.273     0.297    1.79e-4  4.51e-1
#> 11 SYN.EDM     0.0375    0.0349    0.117     0.113     0.303   -6.35e-4  5.89e-1
#> 12 C5.EDM      0.00207   0.000956  0.0926    0.0962    0.214   -1.37e-4  3.16e-1
#> 13 DUB.EFS    -0.0331   -0.0272    0.0586    0.0515    0.0940   1.72e-4  1.20e-1
#> 14 WTI.MEH    -0.000706  0.00268   0.0981    0.114     0.123    3.93e-4  1.99e-1
#> 15 WTI.MID     0.0207    0.0212    0.0885    0.0963    0.124   -7.62e-4  1.71e-1
#> # … with 8 more variables: Mars.CLO01 <dbl>, Mars.CLO02 <dbl>, LLS.SJA <dbl>,
#> #   SYN.EDM <dbl>, C5.EDM <dbl>, DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>
#> 
#> $sims
#> # A tibble: 10 × 15
#>    WTI.CMA01 WTI.CMA02 WCS.CUS  WCS.HOU WCS.HDY   CL.EDM MSW.EDM Mars.…¹ Mars.…²
#>        <dbl>     <dbl>   <dbl>    <dbl>   <dbl>    <dbl>   <dbl>   <dbl>   <dbl>
#>  1     2.30      2.27  -0.306  -0.0903  -0.473  -0.0560   -0.584 -0.551   0.0561
#>  2     3.04      2.64   0.445   0.279    2.31   -0.104     1.32  -0.176   0.265 
#>  3    -1.87     -2.35   0.616  -0.108    0.473   0.168    -0.423  0.392   0.326 
#>  4    -0.380    -0.133  0.603  -0.00308 -0.274  -0.0433   -0.591  0.659   0.619 
#>  5     0.475     0.796  1.00    1.20     0.0805 -0.123     0.952 -0.809   0.0766
#>  6    -5.17     -5.77   0.293   0.696   -0.782   0.0552   -1.28   0.0333  0.0755
#>  7     2.64      2.11  -0.229  -0.404    2.15    0.164    -0.356  0.119   0.170 
#>  8     0.157     0.277 -0.525  -0.554   -0.958   0.118    -0.114  0.668   0.0375
#>  9    -0.851    -0.530  0.324  -0.212    0.499  -0.00614   1.01   0.572   0.811 
#> 10     0.855     0.372 -0.0825 -0.111    0.525   0.0324    0.814 -0.301   0.134 
#> # … with 6 more variables: LLS.SJA <dbl>, SYN.EDM <dbl>, C5.EDM <dbl>,
#> #   DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>, and abbreviated variable names
#> #   ¹​Mars.CLO01, ²​Mars.CLO02
#>