Cox-Ross-Rubinstein binomial option model

European option binomial model on a stock without dividends.For academic purpose only. Use fOptions::CRRBinomialTreeOptions for real-life usage.

CRReuro(S, X, sigma, r, T2M, N, type)

Arguments

S

Stock price.

X

Strike price.

sigma

Implied volatility e.g. 0.20

r

Risk-free rate.

T2M

Time to maturity in years

N

Number of time steps. Internally dt = T2M/N.

type

"call" or "put"

Value

List of asset price tree, option value tree and option price.

Author

Philippe Cote

Examples

CRReuro(S = 100, X = 100, sigma = 0.2, r = 0.1, T2M = 1, N = 5, type = "call")
#> $asset
#>           [,1]      [,2]      [,3]     [,4]     [,5]     [,6]
#> [1,] 100.00000   0.00000   0.00000   0.0000   0.0000   0.0000
#> [2,]  91.44406 109.35647   0.00000   0.0000   0.0000   0.0000
#> [3,]  83.62017 100.00000 119.58837   0.0000   0.0000   0.0000
#> [4,]  76.46568  91.44406 109.35647 130.7776   0.0000   0.0000
#> [5,]  69.92333  83.62017 100.00000 119.5884 143.0138   0.0000
#> [6,]  63.94073  76.46568  91.44406 109.3565 130.7776 156.3948
#> 
#> $option
#>           [,1]      [,2]      [,3]      [,4]     [,5]     [,6]
#> [1,] 13.515415  0.000000  0.000000  0.000000  0.00000  0.00000
#> [2,]  6.365307 18.937675  0.000000  0.000000  0.00000  0.00000
#> [3,]  1.813700  9.740419 25.965507  0.000000  0.00000  0.00000
#> [4,]  0.000000  3.133870 14.656468 34.698679  0.00000  0.00000
#> [5,]  0.000000  0.000000  5.414976 21.568506 44.99392  0.00000
#> [6,]  0.000000  0.000000  0.000000  9.356469 30.77762 56.39483
#> 
#> $price
#> [1] 13.51541
#> 
#> $note
#> [1] "ok"
#>