Markowitz Efficient Frontier

Generates random portfolio weights statistics based on absolute returns. Note returns are neither in percentages nor annualized. This function is used for commodities where returns are dollars per units for real assets e.g. storage tanks, pipelines...

efficientFrontier(
  nsims = 5000,
  x = RTL::fizdiffs %>% dplyr::select(date, dplyr::contains("WCS")),
  expectedReturns = NULL
)

Arguments

nsims

Number of portfolio simulations. Defaults to 5000

x

List as provided by output of RTL::simMultivariates().

expectedReturns

Defaults to NULL using periodic returns means.

Value

List of portfolios and chart of efficient frontier

Author

Philippe Cote

Examples

x =  RTL::fizdiffs %>% dplyr::select(date, dplyr::contains("WCS"))
efficientFrontier(nsims = 10, x = x, expectedReturns = NULL)
#> $data
#> # A tibble: 13 × 5
#>        Return  Risk SharpeRatio data             desc                           
#>         <dbl> <dbl>       <dbl> <list>           <chr>                          
#>  1 -0.000251  0.463   -0.000541 <tibble [1 × 3]> WCS.CUS = 1, WCS.HOU = 0, WCS.…
#>  2 -0.0000862 0.482   -0.000179 <tibble [1 × 3]> WCS.CUS = 0, WCS.HOU = 1, WCS.…
#>  3  0.00128   1.06     0.00120  <tibble [1 × 3]> WCS.CUS = 0, WCS.HOU = 0, WCS.…
#>  4  0.000821  0.766    0.00107  <tibble [1 × 3]> WCS.CUS = 0.21, WCS.HOU = 0.09…
#>  5  0.000164  0.446    0.000367 <tibble [1 × 3]> WCS.CUS = 0.322, WCS.HOU = 0.4…
#>  6 -0.000123  0.435   -0.000282 <tibble [1 × 3]> WCS.CUS = 0.589, WCS.HOU = 0.3…
#>  7  0.0000488 0.431    0.000113 <tibble [1 × 3]> WCS.CUS = 0.415, WCS.HOU = 0.4…
#>  8  0.000621  0.636    0.000977 <tibble [1 × 3]> WCS.CUS = 0.192, WCS.HOU = 0.2…
#>  9  0.000512  0.574    0.000891 <tibble [1 × 3]> WCS.CUS = 0.188, WCS.HOU = 0.3…
#> 10  0.000420  0.528    0.000797 <tibble [1 × 3]> WCS.CUS = 0.164, WCS.HOU = 0.4…
#> 11 -0.0000972 0.460   -0.000211 <tibble [1 × 3]> WCS.CUS = 0.174, WCS.HOU = 0.8…
#> 12  0.000105  0.438    0.000241 <tibble [1 × 3]> WCS.CUS = 0.453, WCS.HOU = 0.3…
#> 13  0.0000711 0.436    0.000163 <tibble [1 × 3]> WCS.CUS = 0.253, WCS.HOU = 0.6…
#> 
#> $plot
#> 
efficientFrontier(nsims = 10, x = x, expectedReturns = c(0.5,0.8,0.9))
#> $data
#> # A tibble: 13 × 5
#>    Return  Risk SharpeRatio data             desc                               
#>     <dbl> <dbl>       <dbl> <list>           <chr>                              
#>  1  0.5   0.463       1.08  <tibble [1 × 3]> WCS.CUS = 1, WCS.HOU = 0, WCS.HDY …
#>  2  0.8   0.482       1.66  <tibble [1 × 3]> WCS.CUS = 0, WCS.HOU = 1, WCS.HDY …
#>  3  0.9   1.06        0.846 <tibble [1 × 3]> WCS.CUS = 0, WCS.HOU = 0, WCS.HDY …
#>  4  0.725 0.553       1.31  <tibble [1 × 3]> WCS.CUS = 0.394, WCS.HOU = 0.175, …
#>  5  0.768 0.486       1.58  <tibble [1 × 3]> WCS.CUS = 0.213, WCS.HOU = 0.468, …
#>  6  0.527 0.452       1.17  <tibble [1 × 3]> WCS.CUS = 0.92, WCS.HOU = 0.055, W…
#>  7  0.761 0.512       1.49  <tibble [1 × 3]> WCS.CUS = 0.254, WCS.HOU = 0.378, …
#>  8  0.712 0.500       1.42  <tibble [1 × 3]> WCS.CUS = 0.409, WCS.HOU = 0.243, …
#>  9  0.860 0.750       1.15  <tibble [1 × 3]> WCS.CUS = 0.023, WCS.HOU = 0.303, …
#> 10  0.660 0.428       1.54  <tibble [1 × 3]> WCS.CUS = 0.508, WCS.HOU = 0.369, …
#> 11  0.625 0.431       1.45  <tibble [1 × 3]> WCS.CUS = 0.628, WCS.HOU = 0.237, …
#> 12  0.677 0.460       1.47  <tibble [1 × 3]> WCS.CUS = 0.498, WCS.HOU = 0.237, …
#> 13  0.685 0.430       1.59  <tibble [1 × 3]> WCS.CUS = 0.419, WCS.HOU = 0.478, …
#> 
#> $plot
#>