Markowitz Efficient Frontier

Generates random portfolio weights statistics based on absolute returns.

efficientFrontier(
  nsims = 5000,
  x = RTL::fizdiffs %>% dplyr::select(date, dplyr::contains("WCS")),
  expectedReturns = NULL
)

Arguments

nsims

Number of portfolio simulations. Defaults to 5000

x

List as provided by output of RTL::simMultivariates().

expectedReturns

Defaults to NULL using periodic returns means.

Value

List of portfolios and chart of efficient frontier

Details

Commodities

Unlike traditional portfolio management, in commodities many transactions are with derivatives (futures and swaps) and have zero or low initial investments.

Return types

This function is used for commodities where returns are dollars per units for real assets e.g. storage tanks, pipelines...Here we measure directly the periodic return in dollars per contract unit.

Empirical Finance

I would encourage you to pick a commodity futures contract of your choice and draw a scatter plot of price level versus the daily dollar per unit change as measure of risk. As a trading analyst or risk manager, then ask yourself about the implications of using log returns that you then re-apply to current forward curve level to arrive at a dollar risk measure per units instead of measuring directly risk in dollars per unit.

Author

Philippe Cote

Examples

x =  RTL::fizdiffs %>% dplyr::select(date, dplyr::contains("WCS"))
efficientFrontier(nsims = 10, x = x, expectedReturns = NULL)
#> $data
#> # A tibble: 13 × 5
#>        Return  Risk SharpeRatio data             desc                           
#>         <dbl> <dbl>       <dbl> <list>           <chr>                          
#>  1 -0.000251  0.463   -0.000541 <tibble [1 × 3]> WCS.CUS = 1, WCS.HOU = 0, WCS.…
#>  2 -0.0000862 0.482   -0.000179 <tibble [1 × 3]> WCS.CUS = 0, WCS.HOU = 1, WCS.…
#>  3  0.00128   1.06     0.00120  <tibble [1 × 3]> WCS.CUS = 0, WCS.HOU = 0, WCS.…
#>  4  0.000346  0.496    0.000699 <tibble [1 × 3]> WCS.CUS = 0.1, WCS.HOU = 0.571…
#>  5 -0.000117  0.440   -0.000266 <tibble [1 × 3]> WCS.CUS = 0.842, WCS.HOU = 0.0…
#>  6  0.000546  0.595    0.000918 <tibble [1 × 3]> WCS.CUS = 0.214, WCS.HOU = 0.2…
#>  7 -0.000144  0.448   -0.000321 <tibble [1 × 3]> WCS.CUS = 0.406, WCS.HOU = 0.5…
#>  8  0.000416  0.537    0.000776 <tibble [1 × 3]> WCS.CUS = 0.328, WCS.HOU = 0.2…
#>  9 -0.0000455 0.433   -0.000105 <tibble [1 × 3]> WCS.CUS = 0.712, WCS.HOU = 0.1…
#> 10  0.000341  0.497    0.000686 <tibble [1 × 3]> WCS.CUS = 0.226, WCS.HOU = 0.4…
#> 11  0.000328  0.492    0.000667 <tibble [1 × 3]> WCS.CUS = 0.218, WCS.HOU = 0.4…
#> 12  0.000535  0.592    0.000904 <tibble [1 × 3]> WCS.CUS = 0.25, WCS.HOU = 0.26…
#> 13  0.000649  0.648    0.00100  <tibble [1 × 3]> WCS.CUS = 0.136, WCS.HOU = 0.3…
#> 
#> $plot
#> 
efficientFrontier(nsims = 10, x = x, expectedReturns = c(0.5,0.8,0.9))
#> $data
#> # A tibble: 13 × 5
#>    Return  Risk SharpeRatio data             desc                               
#>     <dbl> <dbl>       <dbl> <list>           <chr>                              
#>  1  0.5   0.463       1.08  <tibble [1 × 3]> WCS.CUS = 1, WCS.HOU = 0, WCS.HDY …
#>  2  0.8   0.482       1.66  <tibble [1 × 3]> WCS.CUS = 0, WCS.HOU = 1, WCS.HDY …
#>  3  0.9   1.06        0.846 <tibble [1 × 3]> WCS.CUS = 0, WCS.HOU = 0, WCS.HDY …
#>  4  0.694 0.440       1.58  <tibble [1 × 3]> WCS.CUS = 0.422, WCS.HOU = 0.371, …
#>  5  0.802 0.505       1.59  <tibble [1 × 3]> WCS.CUS = 0.11, WCS.HOU = 0.543, W…
#>  6  0.761 0.536       1.42  <tibble [1 × 3]> WCS.CUS = 0.265, WCS.HOU = 0.327, …
#>  7  0.747 0.653       1.15  <tibble [1 × 3]> WCS.CUS = 0.362, WCS.HOU = 0.079, …
#>  8  0.785 0.521       1.51  <tibble [1 × 3]> WCS.CUS = 0.177, WCS.HOU = 0.443, …
#>  9  0.802 0.625       1.28  <tibble [1 × 3]> WCS.CUS = 0.171, WCS.HOU = 0.302, …
#> 10  0.847 0.931       0.910 <tibble [1 × 3]> WCS.CUS = 0.13, WCS.HOU = 0.004, W…
#> 11  0.725 0.502       1.45  <tibble [1 × 3]> WCS.CUS = 0.367, WCS.HOU = 0.281, …
#> 12  0.588 0.434       1.35  <tibble [1 × 3]> WCS.CUS = 0.737, WCS.HOU = 0.174, …
#> 13  0.703 0.448       1.57  <tibble [1 × 3]> WCS.CUS = 0.403, WCS.HOU = 0.364, …
#> 
#> $plot
#>