returns

Computes periodic returns from a dataframe ordered by date

returns(df = dflong, retType = "abs", period.return = 1, spread = FALSE)

Arguments

df
Long dataframe with colnames = c("date","value","series")

retType
"abs" for absolute, "rel" for relative, or "log" for log returns.

period.return
Number of rows over which to compute returns.

spread
TRUE if you want to spread into a long dataframe.

Value

A dataframe object of returns.

Author

Philippe Cote

Examples

returns(df = dflong,retType = "abs",period.return = 1,spread = TRUE)

#> # A tibble: 1,511 x 145
#>    date         BRN01   BRN02   BRN03  BRN04 BRN05 BRN06  BRN07  BRN08  BRN09
#>    <date>       <dbl>   <dbl>   <dbl>  <dbl> <dbl> <dbl>  <dbl>  <dbl>  <dbl>
#>  1 2015-01-05 -3.31   -3.09   -2.98   -2.91  -2.82 -2.7  -2.57  -2.48  -2.40 
#>  2 2015-01-06 -2.01   -2.04   -2.04   -2.04  -2.05 -2.03 -2     -1.96  -1.94 
#>  3 2015-01-07  0.0500  0.0400  0.0400  0     -0.05 -0.13 -0.21  -0.28  -0.34 
#>  4 2015-01-08 -0.190  -0.14   -0.06    0      0.05  0.1   0.14   0.160  0.190
#>  5 2015-01-09 -0.85   -0.73   -0.63   -0.52  -0.43 -0.35 -0.28  -0.220 -0.170
#>  6 2015-01-12 -2.68   -2.60   -2.49   -2.39  -2.31 -2.23 -2.18  -2.15  -2.14 
#>  7 2015-01-13 -0.840  -0.88   -0.89   -0.880 -0.88 -0.85 -0.800 -0.74  -0.69 
#>  8 2015-01-14  2.10    2.04    1.94    1.86   1.81  1.75  1.69   1.62   1.55 
#>  9 2015-01-15 -1.02   -1.59   -1.55   -1.51  -1.48 -1.48 -1.51  -1.54  -1.58 
#> 10 2015-01-16  2.5     3.20    3.20    3.11   2.97  2.73  2.48   2.33   2.25 
#> # … with 1,501 more rows, and 135 more variables: BRN10 <dbl>, BRN11 <dbl>,
#> #   BRN12 <dbl>, BRN13 <dbl>, BRN14 <dbl>, BRN15 <dbl>, BRN16 <dbl>,
#> #   BRN17 <dbl>, BRN18 <dbl>, BRN19 <dbl>, BRN20 <dbl>, BRN21 <dbl>,
#> #   BRN22 <dbl>, BRN23 <dbl>, BRN24 <dbl>, BRN25 <dbl>, BRN26 <dbl>,
#> #   BRN27 <dbl>, BRN28 <dbl>, BRN29 <dbl>, BRN30 <dbl>, BRN31 <dbl>,
#> #   BRN32 <dbl>, BRN33 <dbl>, BRN34 <dbl>, BRN35 <dbl>, BRN36 <dbl>,
#> #   CL01 <dbl>, CL02 <dbl>, CL03 <dbl>, CL04 <dbl>, CL05 <dbl>, CL06 <dbl>,
#> #   CL07 <dbl>, CL08 <dbl>, CL09 <dbl>, CL10 <dbl>, CL11 <dbl>, CL12 <dbl>,
#> #   CL13 <dbl>, CL14 <dbl>, CL15 <dbl>, CL16 <dbl>, CL17 <dbl>, CL18 <dbl>,
#> #   CL19 <dbl>, CL20 <dbl>, CL21 <dbl>, CL22 <dbl>, CL23 <dbl>, CL24 <dbl>,
#> #   CL25 <dbl>, CL26 <dbl>, CL27 <dbl>, CL28 <dbl>, CL29 <dbl>, CL30 <dbl>,
#> #   CL31 <dbl>, CL32 <dbl>, CL33 <dbl>, CL34 <dbl>, CL35 <dbl>, CL36 <dbl>,
#> #   HO01 <dbl>, HO02 <dbl>, HO03 <dbl>, HO04 <dbl>, HO05 <dbl>, HO06 <dbl>,
#> #   HO07 <dbl>, HO08 <dbl>, HO09 <dbl>, HO10 <dbl>, HO11 <dbl>, HO12 <dbl>,
#> #   HO13 <dbl>, HO14 <dbl>, HO15 <dbl>, HO16 <dbl>, HO17 <dbl>, HO18 <dbl>,
#> #   NG01 <dbl>, NG02 <dbl>, NG03 <dbl>, NG04 <dbl>, NG05 <dbl>, NG06 <dbl>,
#> #   NG07 <dbl>, NG08 <dbl>, NG09 <dbl>, NG10 <dbl>, NG11 <dbl>, NG12 <dbl>,
#> #   NG13 <dbl>, NG14 <dbl>, NG15 <dbl>, NG16 <dbl>, NG17 <dbl>, NG18 <dbl>,
#> #   NG19 <dbl>, …

returns(df = dflong,retType = "abs",period.return = 1,spread = FALSE)

#> # A tibble: 213,659 x 4
#> # Groups:   series [144]
#>    date       series value returns
#>    <date>     <chr>  <dbl>   <dbl>
#>  1 2015-01-05 CL01    50.0   -2.65
#>  2 2015-01-05 CL02    50.5   -2.59
#>  3 2015-01-05 CL03    51.2   -2.52
#>  4 2015-01-05 CL04    52.0   -2.47
#>  5 2015-01-05 CL05    52.7   -2.42
#>  6 2015-01-05 CL06    53.5   -2.34
#>  7 2015-01-05 CL07    54.2   -2.25
#>  8 2015-01-05 CL08    54.9   -2.18
#>  9 2015-01-05 CL09    55.6   -2.11
#> 10 2015-01-05 CL10    56.3   -2.04
#> # … with 213,649 more rows