Compute absolute, relative or log returns.

Computes periodic returns from a dataframe ordered by date

returns(df = dflong, retType = "abs", period.return = 1, spread = FALSE)

Arguments

df

Long dataframe with colnames = c("date","value","series")

retType

"abs" for absolute, "rel" for relative, or "log" for log returns.

period.return

Number of rows over which to compute returns.

spread

TRUE if you want to spread into a long dataframe.

Value

A dataframe object of returns.

Author

Philippe Cote

Examples

x <- dflong %>% dplyr::filter(grepl("CL01", series))
returns(df = x, retType = "abs", period.return = 1, spread = TRUE)
#> # A tibble: 4,747 × 2
#>    date          CL01
#>    <date>       <dbl>
#>  1 2004-01-06 -0.0800
#>  2 2004-01-07 -0.0800
#>  3 2004-01-08  0.360 
#>  4 2004-01-09  0.330 
#>  5 2004-01-12  0.410 
#>  6 2004-01-13 -0.290 
#>  7 2004-01-14  0.0700
#>  8 2004-01-15 -1.06  
#>  9 2004-01-16  1.63  
#> 10 2004-01-20  1.13  
#> # … with 4,737 more rows