rolladjust

Returns a xts price or return object adjusted for contract roll. The methodology used to adjust returns is to remove the daily returns on the day after expiry and for prices to adjust historical rolling front month contracts by the size of the roll at each expiry. This is conducive to quantitative trading strategies as it reflects the PL of a financial trader.

rolladjust(x, commodityname = c("cmewti"), rolltype = c("Last.Trade"), ...)

Arguments

x
An xts object of prices or returns.

commodityname
Name of commodity in expiry_table. See example below for values.

rolltype
Type of contract roll: "Last.Trade" or "First.Notice".

...
Other parms

Value

Roll-adjusted xts object of returns

Author

Philippe Cote

Examples

unique(expiry_table$cmdty) # for list of commodity names

#>  [1] "cmeng"        "cmerbob"      "cmeulsd"      "cmewti"       "comexgold"   
#>  [6] "comexsilver"  "icebrent"     "icegasoil"    "lmealum"      "lmecopper"   
#> [11] "lmelead"      "lmenickel"    "lmetin"       "lmezinc"      "PJM.WH.RT.OP"
#> [16] "PJM.WH.RT.P" 

ret <- returns(df=dflong,retType="abs",period.return=1,spread=TRUE)[,1:2] 
rolladjust(x=ret,commodityname=c("cmewti"),rolltype=c("Last.Trade"))

#> # A tibble: 1,440 x 2
#>    date         BRN01
#>    <date>       <dbl>
#>  1 2015-01-06 -2.01  
#>  2 2015-01-07  0.0500
#>  3 2015-01-08 -0.190 
#>  4 2015-01-09 -0.85  
#>  5 2015-01-12 -2.68  
#>  6 2015-01-13 -0.840 
#>  7 2015-01-14  2.10  
#>  8 2015-01-15 -1.02  
#>  9 2015-01-16  2.5   
#> 10 2015-01-19 -1.33  
#> # … with 1,430 more rows