Multivariate Normal from historical dataset

Generates multivariate random epsilons using absolute returns.

simMultivariates(nsims = 10, x, s0 = NULL)

Arguments

nsims

Number of simulations. Defaults to 10

x

Wide data frame of prices with date as first column.

s0

Vector of starting value for each variables. Defaults to NULL with zero.

Value

List of means, sds, covariance matrix, correlation matrix and simulated values

Author

Philippe Cote

Examples

simMultivariates(nsims = 10, x = RTL::fizdiffs, s0 = NULL)
#> $ave
#> # A tibble: 1 × 15
#>   WTI.CMA01 WTI.CMA02   WCS.CUS    WCS.HOU WCS.HDY     CL.EDM MSW.EDM Mars.CLO01
#>       <dbl>     <dbl>     <dbl>      <dbl>   <dbl>      <dbl>   <dbl>      <dbl>
#> 1    0.0432    0.0403 -0.000251 -0.0000862 0.00128 -0.0000392 0.00165  0.0000392
#> # … with 7 more variables: Mars.CLO02 <dbl>, LLS.SJA <dbl>, SYN.EDM <dbl>,
#> #   C5.EDM <dbl>, DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>
#> 
#> $sd
#> # A tibble: 1 × 15
#>   WTI.CMA01 WTI.CMA02 WCS.CUS WCS.HOU WCS.HDY CL.EDM MSW.EDM Mars.CLO01
#>       <dbl>     <dbl>   <dbl>   <dbl>   <dbl>  <dbl>   <dbl>      <dbl>
#> 1      1.55      1.49   0.463   0.482    1.06 0.0709   0.889      0.689
#> # … with 7 more variables: Mars.CLO02 <dbl>, LLS.SJA <dbl>, SYN.EDM <dbl>,
#> #   C5.EDM <dbl>, DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>
#> 
#> $corMat
#>               WTI.CMA01    WTI.CMA02      WCS.CUS     WCS.HOU      WCS.HDY
#> WTI.CMA01   1.000000000  0.996663551  0.001935236 0.016042884 -0.063293605
#> WTI.CMA02   0.996663551  1.000000000 -0.009006441 0.008382594 -0.063819131
#> WCS.CUS     0.001935236 -0.009006441  1.000000000 0.790344972  0.134270690
#> WCS.HOU     0.016042884  0.008382594  0.790344972 1.000000000  0.107504073
#> WCS.HDY    -0.063293605 -0.063819131  0.134270690 0.107504073  1.000000000
#> CL.EDM      0.041053825  0.038211835  0.027433507 0.026835201  0.002880414
#> MSW.EDM    -0.019722790 -0.020646399  0.148194720 0.133516377  0.372028411
#> Mars.CLO01 -0.057586763 -0.043672374  0.290870653 0.349281025  0.074653515
#> Mars.CLO02  0.002201803  0.005265953  0.413220470 0.501805860  0.099634679
#> LLS.SJA     0.044072433  0.051583201  0.167042946 0.186964025 -0.072199226
#> SYN.EDM     0.025598008  0.025249698  0.174715532 0.167273516  0.362052732
#> C5.EDM      0.027468505  0.021129184  0.143636143 0.160006245  0.236821355
#> DUB.EFS    -0.071572448 -0.059165937  0.245490787 0.194429000  0.099460734
#> WTI.MEH     0.018215840  0.030093599  0.274851205 0.333309836 -0.018885511
#> WTI.MID     0.026156877  0.036578180  0.264079587 0.277074729  0.088077406
#>                   CL.EDM       MSW.EDM  Mars.CLO01  Mars.CLO02     LLS.SJA
#> WTI.CMA01   0.0410538245 -0.0197227905 -0.05758676 0.002201803  0.04407243
#> WTI.CMA02   0.0382118354 -0.0206463991 -0.04367237 0.005265953  0.05158320
#> WCS.CUS     0.0274335066  0.1481947199  0.29087065 0.413220470  0.16704295
#> WCS.HOU     0.0268352007  0.1335163772  0.34928102 0.501805860  0.18696402
#> WCS.HDY     0.0028804135  0.3720284108  0.07465351 0.099634679 -0.07219923
#> CL.EDM      1.0000000000 -0.0002566291  0.00927482 0.009048434  0.01440119
#> MSW.EDM    -0.0002566291  1.0000000000  0.07720471 0.134135452 -0.04410322
#> Mars.CLO01  0.0092748202  0.0772047139  1.00000000 0.560030780  0.38144604
#> Mars.CLO02  0.0090484345  0.1341354516  0.56003078 1.000000000  0.09146071
#> LLS.SJA     0.0144011894 -0.0441032202  0.38144604 0.091460712  1.00000000
#> SYN.EDM    -0.0035381299  0.7174625480  0.08706154 0.147056199 -0.04616890
#> C5.EDM      0.0387071942  0.3641947093  0.06390083 0.102275236  0.04523336
#> DUB.EFS     0.0278091255  0.0780346169  0.16237680 0.287707795  0.05770498
#> WTI.MEH     0.0069449988  0.0139026558  0.63703333 0.350991458  0.82703145
#> WTI.MID     0.0140235398  0.0508530553  0.51105994 0.294828994  0.55094849
#>                SYN.EDM     C5.EDM     DUB.EFS      WTI.MEH    WTI.MID
#> WTI.CMA01   0.02559801 0.02746851 -0.07157245  0.018215840 0.02615688
#> WTI.CMA02   0.02524970 0.02112918 -0.05916594  0.030093599 0.03657818
#> WCS.CUS     0.17471553 0.14363614  0.24549079  0.274851205 0.26407959
#> WCS.HOU     0.16727352 0.16000624  0.19442900  0.333309836 0.27707473
#> WCS.HDY     0.36205273 0.23682135  0.09946073 -0.018885511 0.08807741
#> CL.EDM     -0.00353813 0.03870719  0.02780913  0.006944999 0.01402354
#> MSW.EDM     0.71746255 0.36419471  0.07803462  0.013902656 0.05085306
#> Mars.CLO01  0.08706154 0.06390083  0.16237680  0.637033332 0.51105994
#> Mars.CLO02  0.14705620 0.10227524  0.28770779  0.350991458 0.29482899
#> LLS.SJA    -0.04616890 0.04523336  0.05770498  0.827031446 0.55094849
#> SYN.EDM     1.00000000 0.36956557  0.10108813  0.012647144 0.05904903
#> C5.EDM      0.36956557 1.00000000  0.08753485  0.095284198 0.07535497
#> DUB.EFS     0.10108813 0.08753485  1.00000000  0.118223113 0.13628714
#> WTI.MEH     0.01264714 0.09528420  0.11822311  1.000000000 0.67778855
#> WTI.MID     0.05904903 0.07535497  0.13628714  0.677788552 1.00000000
#> 
#> $coVaR
#> # A tibble: 15 × 16
#>    series     WTI.CMA01 WTI.CMA02   WCS.CUS  WCS.HOU   WCS.HDY   CL.EDM  MSW.EDM
#>    <chr>          <dbl>     <dbl>     <dbl>    <dbl>     <dbl>    <dbl>    <dbl>
#>  1 WTI.CMA01    2.40      2.29     0.00139  0.0120   -0.104     4.51e-3 -2.71e-2
#>  2 WTI.CMA02    2.29      2.21    -0.00620  0.00600  -0.101     4.03e-3 -2.73e-2
#>  3 WCS.CUS      0.00139  -0.00620  0.215    0.176     0.0662    9.01e-4  6.10e-2
#>  4 WCS.HOU      0.0120    0.00600  0.176    0.232     0.0551    9.17e-4  5.72e-2
#>  5 WCS.HDY     -0.104    -0.101    0.0662   0.0551    1.13      2.17e-4  3.52e-1
#>  6 CL.EDM       0.00451   0.00403  0.000901 0.000917  0.000217  5.03e-3 -1.62e-5
#>  7 MSW.EDM     -0.0271   -0.0273   0.0610   0.0572    0.352    -1.62e-5  7.90e-1
#>  8 Mars.CLO01  -0.0615   -0.0447   0.0928   0.116     0.0547    4.53e-4  4.73e-2
#>  9 Mars.CLO02   0.00131   0.00301  0.0736   0.0930    0.0408    2.47e-4  4.58e-2
#> 10 LLS.SJA      0.123     0.139    0.140    0.163    -0.139     1.85e-3 -7.09e-2
#> 11 SYN.EDM      0.0366    0.0346   0.0747   0.0744    0.356    -2.32e-4  5.89e-1
#> 12 C5.EDM       0.0319    0.0235   0.0499   0.0578    0.189     2.06e-3  2.43e-1
#> 13 DUB.EFS     -0.0473   -0.0375   0.0485   0.0400    0.0452    8.42e-4  2.96e-2
#> 14 WTI.MEH      0.0194    0.0307   0.0874   0.110    -0.0138    3.38e-4  8.48e-3
#> 15 WTI.MID      0.0293    0.0393   0.0885   0.0966    0.0678    7.20e-4  3.27e-2
#> # … with 8 more variables: Mars.CLO01 <dbl>, Mars.CLO02 <dbl>, LLS.SJA <dbl>,
#> #   SYN.EDM <dbl>, C5.EDM <dbl>, DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>
#> 
#> $sims
#> # A tibble: 10 × 15
#>    WTI.CMA01 WTI.CMA02 WCS.CUS WCS.HOU WCS.HDY   CL.EDM  MSW.EDM Mars.CLO01
#>        <dbl>     <dbl>   <dbl>   <dbl>   <dbl>    <dbl>    <dbl>      <dbl>
#>  1    0.105     0.202  -0.785  -0.532  -0.136   0.0937  -0.693       -0.799
#>  2   -0.0267    0.0870 -0.349  -0.764  -2.29    0.147   -1.67        -0.387
#>  3   -2.25     -2.27    0.580   0.405   0.359   0.0613  -0.00531     -0.428
#>  4    0.855     0.705  -0.558  -0.336  -0.624   0.0475  -0.428       -0.715
#>  5    2.30      2.13    0.189  -0.746   0.266   0.0648  -0.434        0.486
#>  6   -0.927    -0.872  -0.0138  0.0945 -0.428   0.0767   0.482       -0.736
#>  7    0.144     0.139  -0.289  -0.224  -0.612  -0.0653   0.729        0.371
#>  8    0.391     0.228  -0.0178 -0.0421 -0.304   0.00643  0.684       -1.66 
#>  9    1.88      1.76    0.0851  0.352  -0.355   0.0197  -0.114        0.203
#> 10   -1.59     -1.51    0.140   0.207   0.0956  0.0407   0.872       -0.340
#> # … with 7 more variables: Mars.CLO02 <dbl>, LLS.SJA <dbl>, SYN.EDM <dbl>,
#> #   C5.EDM <dbl>, DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>
#>