Multivariate Normal from historical dataset

Generates multivariate random epsilons using absolute returns.

simMultivariates(nsims = 10, x, s0 = NULL)

Arguments

nsims

Number of simulations. Defaults to 10

x

Wide data frame of prices with date as first column.

s0

Vector of starting value for each variables. Defaults to NULL with zero.

Value

List of means, sds, covariance matrix, correlation matrix and simulated values

Author

Philippe Cote

Examples

simMultivariates(nsims = 10, x = RTL::fizdiffs, s0 = NULL)
#> $ave
#> # A tibble: 1 × 15
#>   WTI.CMA01 WTI.CMA02   WCS.CUS  WCS.HOU  WCS.HDY     CL.EDM  MSW.EDM Mars.CLO01
#>       <dbl>     <dbl>     <dbl>    <dbl>    <dbl>      <dbl>    <dbl>      <dbl>
#> 1    0.0469    0.0444 -0.000910 -0.00110 -0.00253 -0.0000154 -4.86e-4   -0.00319
#> # … with 7 more variables: Mars.CLO02 <dbl>, LLS.SJA <dbl>, SYN.EDM <dbl>,
#> #   C5.EDM <dbl>, DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>
#> 
#> $sd
#> # A tibble: 1 × 15
#>   WTI.CMA01 WTI.CMA02 WCS.CUS WCS.HOU WCS.HDY CL.EDM MSW.EDM Mars.CLO01
#>       <dbl>     <dbl>   <dbl>   <dbl>   <dbl>  <dbl>   <dbl>      <dbl>
#> 1      1.55      1.49   0.464   0.479    1.08 0.0719   0.891      0.684
#> # … with 7 more variables: Mars.CLO02 <dbl>, LLS.SJA <dbl>, SYN.EDM <dbl>,
#> #   C5.EDM <dbl>, DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>
#> 
#> $corMat
#>              WTI.CMA01    WTI.CMA02      WCS.CUS    WCS.HOU     WCS.HDY
#> WTI.CMA01   1.00000000  0.996616358  0.010780933 0.02549138 -0.05836191
#> WTI.CMA02   0.99661636  1.000000000 -0.002314081 0.01568141 -0.05927448
#> WCS.CUS     0.01078093 -0.002314081  1.000000000 0.79146799  0.15048941
#> WCS.HOU     0.02549138  0.015681412  0.791467993 1.00000000  0.12190066
#> WCS.HDY    -0.05836191 -0.059274484  0.150489411 0.12190066  1.00000000
#> CL.EDM      0.00626971  0.011071672 -0.028980957 0.01198717 -0.02294462
#> MSW.EDM    -0.01131609 -0.013460620  0.158795136 0.14033519  0.38126587
#> Mars.CLO01 -0.04793420 -0.035422794  0.279593960 0.33724062  0.08180934
#> Mars.CLO02  0.01422607  0.014608899  0.403195234 0.48911994  0.12410129
#> LLS.SJA     0.04865681  0.055531089  0.133983745 0.15461914 -0.08181864
#> SYN.EDM     0.03104621  0.029535450  0.169421143 0.15882205  0.36723785
#> C5.EDM      0.03404480  0.025955082  0.155081340 0.16284477  0.24000640
#> DUB.EFS    -0.04579531 -0.036335773  0.247286834 0.19492820  0.14051989
#> WTI.MEH     0.02861866  0.039148894  0.248634936 0.30581201 -0.01186668
#> WTI.MID     0.03550998  0.044733193  0.237368184 0.24844531  0.09619567
#>                  CL.EDM      MSW.EDM  Mars.CLO01  Mars.CLO02     LLS.SJA
#> WTI.CMA01   0.006269710 -0.011316086 -0.04793420 0.014226069  0.04865681
#> WTI.CMA02   0.011071672 -0.013460620 -0.03542279 0.014608899  0.05553109
#> WCS.CUS    -0.028980957  0.158795136  0.27959396 0.403195234  0.13398374
#> WCS.HOU     0.011987168  0.140335191  0.33724062 0.489119941  0.15461914
#> WCS.HDY    -0.022944619  0.381265873  0.08180934 0.124101292 -0.08181864
#> CL.EDM      1.000000000 -0.009852795 -0.01431618 0.002211077 -0.06670005
#> MSW.EDM    -0.009852795  1.000000000  0.07998031 0.142842475 -0.05814979
#> Mars.CLO01 -0.014316183  0.079980312  1.00000000 0.555479907  0.36022304
#> Mars.CLO02  0.002211077  0.142842475  0.55547991 1.000000000  0.04756699
#> LLS.SJA    -0.066700051 -0.058149794  0.36022304 0.047566987  1.00000000
#> SYN.EDM     0.031646690  0.718585843  0.07913249 0.136499864 -0.06535321
#> C5.EDM      0.029790735  0.376895889  0.07022552 0.117128313  0.02901394
#> DUB.EFS    -0.034102297  0.102738364  0.16032401 0.291585726  0.02421842
#> WTI.MEH    -0.033152708  0.011420582  0.62268945 0.316475378  0.81836552
#> WTI.MID    -0.032735447  0.048816917  0.49421608 0.258536171  0.53499865
#>                 SYN.EDM     C5.EDM     DUB.EFS      WTI.MEH     WTI.MID
#> WTI.CMA01   0.031046211 0.03404480 -0.04579531  0.028618656  0.03550998
#> WTI.CMA02   0.029535450 0.02595508 -0.03633577  0.039148894  0.04473319
#> WCS.CUS     0.169421143 0.15508134  0.24728683  0.248634936  0.23736818
#> WCS.HOU     0.158822051 0.16284477  0.19492820  0.305812011  0.24844531
#> WCS.HDY     0.367237852 0.24000640  0.14051989 -0.011866683  0.09619567
#> CL.EDM      0.031646690 0.02979073 -0.03410230 -0.033152708 -0.03273545
#> MSW.EDM     0.718585843 0.37689589  0.10273836  0.011420582  0.04881692
#> Mars.CLO01  0.079132486 0.07022552  0.16032401  0.622689453  0.49421608
#> Mars.CLO02  0.136499864 0.11712831  0.29158573  0.316475378  0.25853617
#> LLS.SJA    -0.065353215 0.02901394  0.02421842  0.818365520  0.53499865
#> SYN.EDM     1.000000000 0.37161681  0.10412300 -0.002107243  0.04594094
#> C5.EDM      0.371616806 1.00000000  0.11401455  0.095197500  0.07523695
#> DUB.EFS     0.104122998 0.11401455  1.00000000  0.102373981  0.11983673
#> WTI.MEH    -0.002107243 0.09519750  0.10237398  1.000000000  0.66642912
#> WTI.MID     0.045940938 0.07523695  0.11983673  0.666429117  1.00000000
#> 
#> $coVaR
#> # A tibble: 15 × 16
#>    series     WTI.CMA01 WTI.CMA02   WCS.CUS  WCS.HOU  WCS.HDY    CL.EDM  MSW.EDM
#>    <chr>          <dbl>     <dbl>     <dbl>    <dbl>    <dbl>     <dbl>    <dbl>
#>  1 WTI.CMA01   2.40       2.30     0.00775  0.0189   -0.0978    6.98e-4 -1.56e-2
#>  2 WTI.CMA02   2.30       2.21    -0.00160  0.0112   -0.0953    1.18e-3 -1.78e-2
#>  3 WCS.CUS     0.00775   -0.00160  0.215    0.176     0.0755   -9.67e-4  6.57e-2
#>  4 WCS.HOU     0.0189     0.0112   0.176    0.229     0.0631    4.12e-4  5.99e-2
#>  5 WCS.HDY    -0.0978    -0.0953   0.0755   0.0631    1.17     -1.78e-3  3.67e-1
#>  6 CL.EDM      0.000698   0.00118 -0.000967 0.000412 -0.00178   5.16e-3 -6.31e-4
#>  7 MSW.EDM    -0.0156    -0.0178   0.0657   0.0599    0.367    -6.31e-4  7.95e-1
#>  8 Mars.CLO01 -0.0508    -0.0360   0.0888   0.111     0.0605   -7.04e-4  4.88e-2
#>  9 Mars.CLO02  0.00835    0.00823  0.0709   0.0887    0.0508    6.02e-5  4.82e-2
#> 10 LLS.SJA     0.134      0.147    0.111    0.132    -0.157    -8.53e-3 -9.22e-2
#> 11 SYN.EDM     0.0441     0.0402   0.0720   0.0697    0.364     2.08e-3  5.87e-1
#> 12 C5.EDM      0.0412     0.0302   0.0563   0.0610    0.203     1.67e-3  2.63e-1
#> 13 DUB.EFS    -0.0312    -0.0238   0.0505   0.0411    0.0669   -1.08e-3  4.03e-2
#> 14 WTI.MEH     0.0297     0.0390   0.0772   0.0980   -0.00859  -1.59e-3  6.81e-3
#> 15 WTI.MID     0.0389     0.0470   0.0778   0.0840    0.0735   -1.66e-3  3.07e-2
#> # … with 8 more variables: Mars.CLO01 <dbl>, Mars.CLO02 <dbl>, LLS.SJA <dbl>,
#> #   SYN.EDM <dbl>, C5.EDM <dbl>, DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>
#> 
#> $sims
#> # A tibble: 10 × 15
#>    WTI.CMA01 WTI.CMA02 WCS.CUS  WCS.HOU WCS.HDY  CL.EDM MSW.EDM Mars.CLO01
#>        <dbl>     <dbl>   <dbl>    <dbl>   <dbl>   <dbl>   <dbl>      <dbl>
#>  1     1.27      1.13   0.476   0.341    -1.85  -0.0855   0.478    -0.123 
#>  2     0.427     0.636 -0.209  -0.336     0.748  0.0102   1.25     -0.0196
#>  3    -0.400    -0.607 -0.270  -0.168    -0.664  0.0330  -1.57      0.936 
#>  4    -1.25     -1.12   0.320   0.104     1.15   0.0788   1.38     -0.411 
#>  5     0.647     0.500 -0.270   0.00535   0.237 -0.0774   0.105    -0.202 
#>  6    -0.885    -0.745 -0.152   0.295    -1.01   0.0485  -0.292    -0.0432
#>  7    -0.629    -0.763  0.0891 -0.0561    1.88  -0.114    2.49     -0.820 
#>  8     0.708     0.590 -0.134   0.113     0.712 -0.0658  -2.36     -0.144 
#>  9     1.28      1.21  -0.398  -0.0634   -2.81   0.0201   0.178     0.797 
#> 10    -1.60     -1.81   0.246   0.0776   -1.04  -0.163   -0.503     0.279 
#> # … with 7 more variables: Mars.CLO02 <dbl>, LLS.SJA <dbl>, SYN.EDM <dbl>,
#> #   C5.EDM <dbl>, DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>
#>