Multivariate Normal from historical dataset

Generates multivariate random epsilons using absolute returns.

simMultivariates(nsims = 10, x, s0 = NULL)

Arguments

nsims

Number of simulations. Defaults to 10

x

Wide data frame of prices with date as first column.

s0

Vector of starting value for each variables. Defaults to NULL with zero.

Value

List of means, sds, covariance matrix, correlation matrix and simulated values

Author

Philippe Cote

Examples

simMultivariates(nsims = 10, x = RTL::fizdiffs, s0 = NULL)
#> $ave
#> # A tibble: 1 × 15
#>   WTI.CM…¹ WTI.C…²  WCS.CUS  WCS.HOU  WCS.HDY   CL.EDM MSW.EDM Mars.C…³ Mars.C…⁴
#>      <dbl>   <dbl>    <dbl>    <dbl>    <dbl>    <dbl>   <dbl>    <dbl>    <dbl>
#> 1   0.0307  0.0284 -0.00123 -8.04e-4 -0.00329 -3.79e-5 1.97e-4 -0.00147 -3.19e-4
#> # … with 6 more variables: LLS.SJA <dbl>, SYN.EDM <dbl>, C5.EDM <dbl>,
#> #   DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>, and abbreviated variable names
#> #   ¹​WTI.CMA01, ²​WTI.CMA02, ³​Mars.CLO01, ⁴​Mars.CLO02
#> 
#> $sd
#> # A tibble: 1 × 15
#>   WTI.C…¹ WTI.C…² WCS.CUS WCS.HOU WCS.HDY CL.EDM MSW.EDM Mars.…³ Mars.…⁴ LLS.SJA
#>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>  <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1    1.61    1.54   0.478   0.501    1.08 0.0719   0.889   0.692   0.383    1.77
#> # … with 5 more variables: SYN.EDM <dbl>, C5.EDM <dbl>, DUB.EFS <dbl>,
#> #   WTI.MEH <dbl>, WTI.MID <dbl>, and abbreviated variable names ¹​WTI.CMA01,
#> #   ²​WTI.CMA02, ³​Mars.CLO01, ⁴​Mars.CLO02
#> 
#> $corMat
#>                WTI.CMA01    WTI.CMA02       WCS.CUS      WCS.HOU     WCS.HDY
#> WTI.CMA01   1.0000000000  0.959483271 -0.0008827808  0.011764250 -0.03155566
#> WTI.CMA02   0.9594832712  1.000000000 -0.0069864686  0.006136637 -0.03418362
#> WCS.CUS    -0.0008827808 -0.006986469  1.0000000000  0.644018422  0.21488030
#> WCS.HOU     0.0117642502  0.006136637  0.6440184216  1.000000000  0.19760656
#> WCS.HDY    -0.0315556573 -0.034183621  0.2148802970  0.197606556  1.00000000
#> CL.EDM      0.0328037582  0.033385542 -0.0103912007 -0.010239382 -0.00446074
#> MSW.EDM     0.0216506245  0.021073332  0.3214784109  0.291963576  0.31009464
#> Mars.CLO01 -0.0091457461 -0.003734512  0.3114345985  0.334601181  0.15723941
#> Mars.CLO02 -0.0110156967 -0.008399487  0.3614191455  0.392693386  0.18637196
#> LLS.SJA     0.0009422968  0.004949711  0.2875728322  0.315461851  0.14884098
#> SYN.EDM     0.0275668808  0.026093127  0.3054875840  0.290473964  0.30791184
#> C5.EDM      0.0192783834  0.016056233  0.2760650252  0.277220229  0.25008737
#> DUB.EFS    -0.0395838892 -0.033608500  0.3077405686  0.271692787  0.20855539
#> WTI.MEH    -0.0015949607  0.001949660  0.3299372355  0.357160901  0.15895008
#> WTI.MID     0.0245495295  0.025026921  0.2895188928  0.302904597  0.16473613
#>                  CL.EDM     MSW.EDM   Mars.CLO01   Mars.CLO02       LLS.SJA
#> WTI.CMA01   0.032803758 0.021650624 -0.009145746 -0.011015697  0.0009422968
#> WTI.CMA02   0.033385542 0.021073332 -0.003734512 -0.008399487  0.0049497106
#> WCS.CUS    -0.010391201 0.321478411  0.311434598  0.361419145  0.2875728322
#> WCS.HOU    -0.010239382 0.291963576  0.334601181  0.392693386  0.3154618508
#> WCS.HDY    -0.004460740 0.310094639  0.157239412  0.186371957  0.1488409787
#> CL.EDM      1.000000000 0.003576912 -0.023481006 -0.011688760 -0.0130048007
#> MSW.EDM     0.003576912 1.000000000  0.267881458  0.334608504  0.2876432094
#> Mars.CLO01 -0.023481006 0.267881458  1.000000000  0.653323176  0.5312598964
#> Mars.CLO02 -0.011688760 0.334608504  0.653323176  1.000000000  0.5319050323
#> LLS.SJA    -0.013004801 0.287643209  0.531259896  0.531905032  1.0000000000
#> SYN.EDM     0.001788682 0.737507472  0.250766602  0.308174038  0.2574599110
#> C5.EDM      0.011251291 0.463738423  0.225801572  0.286171554  0.2664455157
#> DUB.EFS    -0.003747540 0.323161062  0.253153164  0.328342527  0.2563490237
#> WTI.MEH    -0.018904096 0.342365989  0.489778417  0.538164986  0.5570112977
#> WTI.MID    -0.003365961 0.292675168  0.325369373  0.366363727  0.3611482904
#>                SYN.EDM     C5.EDM     DUB.EFS      WTI.MEH      WTI.MID
#> WTI.CMA01  0.027566881 0.01927838 -0.03958389 -0.001594961  0.024549530
#> WTI.CMA02  0.026093127 0.01605623 -0.03360850  0.001949660  0.025026921
#> WCS.CUS    0.305487584 0.27606503  0.30774057  0.329937235  0.289518893
#> WCS.HOU    0.290473964 0.27722023  0.27169279  0.357160901  0.302904597
#> WCS.HDY    0.307911844 0.25008737  0.20855539  0.158950078  0.164736127
#> CL.EDM     0.001788682 0.01125129 -0.00374754 -0.018904096 -0.003365961
#> MSW.EDM    0.737507472 0.46373842  0.32316106  0.342365989  0.292675168
#> Mars.CLO01 0.250766602 0.22580157  0.25315316  0.489778417  0.325369373
#> Mars.CLO02 0.308174038 0.28617155  0.32834253  0.538164986  0.366363727
#> LLS.SJA    0.257459911 0.26644552  0.25634902  0.557011298  0.361148290
#> SYN.EDM    1.000000000 0.43584016  0.29379630  0.320100381  0.276156922
#> C5.EDM     0.435840162 1.00000000  0.30814559  0.317461719  0.267636825
#> DUB.EFS    0.293796297 0.30814559  1.00000000  0.289242789  0.286480664
#> WTI.MEH    0.320100381 0.31746172  0.28924279  1.000000000  0.502894469
#> WTI.MID    0.276156922 0.26763683  0.28648066  0.502894469  1.000000000
#> 
#> $coVaR
#> # A tibble: 15 × 16
#>    series WTI.CM…¹ WTI.CM…²  WCS.CUS  WCS.HOU  WCS.HDY   CL.EDM MSW.EDM Mars.C…³
#>    <chr>     <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>   <dbl>    <dbl>
#>  1 WTI.C…  2.58e+0  2.37    -6.78e-4  9.46e-3 -5.46e-2  3.79e-3 3.09e-2 -0.0102 
#>  2 WTI.C…  2.37e+0  2.38    -5.15e-3  4.74e-3 -5.68e-2  3.70e-3 2.89e-2 -0.00398
#>  3 WCS.C… -6.78e-4 -0.00515  2.29e-1  1.54e-1  1.11e-1 -3.57e-4 1.37e-1  0.103  
#>  4 WCS.H…  9.46e-3  0.00474  1.54e-1  2.51e-1  1.07e-1 -3.69e-4 1.30e-1  0.116  
#>  5 WCS.H… -5.46e-2 -0.0568   1.11e-1  1.07e-1  1.16e+0 -3.45e-4 2.97e-1  0.117  
#>  6 CL.EDM  3.79e-3  0.00370 -3.57e-4 -3.69e-4 -3.45e-4  5.17e-3 2.29e-4 -0.00117
#>  7 MSW.E…  3.09e-2  0.0289   1.37e-1  1.30e-1  2.97e-1  2.29e-4 7.91e-1  0.165  
#>  8 Mars.… -1.02e-2 -0.00398  1.03e-1  1.16e-1  1.17e-1 -1.17e-3 1.65e-1  0.478  
#>  9 Mars.… -6.78e-3 -0.00496  6.62e-2  7.54e-2  7.70e-2 -3.22e-4 1.14e-1  0.173  
#> 10 LLS.S…  2.68e-3  0.0135   2.43e-1  2.80e-1  2.84e-1 -1.66e-3 4.53e-1  0.651  
#> 11 SYN.E…  4.10e-2  0.0373   1.35e-1  1.35e-1  3.07e-1  1.19e-4 6.07e-1  0.161  
#> 12 C5.EDM  2.45e-2  0.0196   1.05e-1  1.10e-1  2.14e-1  6.41e-4 3.27e-1  0.124  
#> 13 DUB.E… -2.79e-2 -0.0227   6.45e-2  5.96e-2  9.85e-2 -1.18e-4 1.26e-1  0.0767 
#> 14 WTI.M… -1.72e-3  0.00201  1.06e-1  1.20e-1  1.15e-1 -9.10e-4 2.04e-1  0.227  
#> 15 WTI.M…  2.81e-2  0.0275   9.88e-2  1.08e-1  1.27e-1 -1.73e-4 1.86e-1  0.161  
#> # … with 7 more variables: Mars.CLO02 <dbl>, LLS.SJA <dbl>, SYN.EDM <dbl>,
#> #   C5.EDM <dbl>, DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>, and abbreviated
#> #   variable names ¹​WTI.CMA01, ²​WTI.CMA02, ³​Mars.CLO01
#> 
#> $sims
#> # A tibble: 10 × 15
#>    WTI.CMA01 WTI.CMA02 WCS.CUS WCS.HOU WCS.HDY   CL.EDM MSW.EDM Mars.C…¹ Mars.…²
#>        <dbl>     <dbl>   <dbl>   <dbl>   <dbl>    <dbl>   <dbl>    <dbl>   <dbl>
#>  1    -1.14    -0.694    0.433  0.623   -2.55  -0.00868 -0.757    0.449   0.322 
#>  2     2.54     1.84    -0.380 -0.892    0.307  0.0357   2.52    -0.525  -0.184 
#>  3     0.816    0.146    0.345  0.214    0.258  0.0662   0.249    0.0264 -0.547 
#>  4    -0.150   -0.231   -0.415 -0.119    1.00   0.0428   0.758    0.0780  0.238 
#>  5     4.15     4.04     0.694 -0.0316   0.729  0.0701  -0.0395  -1.18   -0.213 
#>  6     1.98     1.67    -0.275  0.0612  -1.02  -0.0134  -0.828   -0.0808 -0.137 
#>  7     0.234   -0.0370  -0.636 -0.116    0.395 -0.0660  -0.779    0.510  -0.0429
#>  8    -1.99    -1.08     0.494  0.915   -2.65  -0.114   -0.731   -0.0197  0.125 
#>  9    -0.240    0.550    0.797  0.937    0.668  0.132   -0.0421   1.19    0.642 
#> 10    -0.336   -1.11    -0.105 -0.0784   0.635 -0.0553  -0.368    0.386   0.375 
#> # … with 6 more variables: LLS.SJA <dbl>, SYN.EDM <dbl>, C5.EDM <dbl>,
#> #   DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>, and abbreviated variable names
#> #   ¹​Mars.CLO01, ²​Mars.CLO02
#>