Multivariate Normal from historical dataset

Generates multivariate random epsilons using absolute returns.

simMultivariates(nsims = 10, x, s0 = NULL)

Arguments

nsims

Number of simulations. Defaults to 10

x

Wide data frame of prices with date as first column.

s0

Vector of starting value for each variables. Defaults to NULL with zero.

Value

List of means, sds, covariance matrix, correlation matrix and simulated values

Author

Philippe Cote

Examples

simMultivariates(nsims = 10, x = RTL::fizdiffs, s0 = NULL)
#> $ave
#> # A tibble: 1 × 15
#>   WTI.CMA01 WTI.CMA02   WCS.CUS  WCS.HOU  WCS.HDY     CL.EDM  MSW.EDM Mars.CLO01
#>       <dbl>     <dbl>     <dbl>    <dbl>    <dbl>      <dbl>    <dbl>      <dbl>
#> 1    0.0469    0.0444 -0.000910 -0.00110 -0.00253 -0.0000154 -4.86e-4   -0.00319
#> # … with 7 more variables: Mars.CLO02 <dbl>, LLS.SJA <dbl>, SYN.EDM <dbl>,
#> #   C5.EDM <dbl>, DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>
#> 
#> $sd
#> # A tibble: 1 × 15
#>   WTI.CMA01 WTI.CMA02 WCS.CUS WCS.HOU WCS.HDY CL.EDM MSW.EDM Mars.CLO01
#>       <dbl>     <dbl>   <dbl>   <dbl>   <dbl>  <dbl>   <dbl>      <dbl>
#> 1      1.55      1.49   0.464   0.479    1.08 0.0719   0.891      0.684
#> # … with 7 more variables: Mars.CLO02 <dbl>, LLS.SJA <dbl>, SYN.EDM <dbl>,
#> #   C5.EDM <dbl>, DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>
#> 
#> $corMat
#>                WTI.CMA01    WTI.CMA02     WCS.CUS      WCS.HOU     WCS.HDY
#> WTI.CMA01   1.0000000000  0.959221146 -0.01138268 -0.007847030 -0.04046452
#> WTI.CMA02   0.9592211463  1.000000000 -0.01696116 -0.012554527 -0.04311311
#> WCS.CUS    -0.0113826814 -0.016961158  1.00000000  0.633681065  0.21147710
#> WCS.HOU    -0.0078470297 -0.012554527  0.63368106  1.000000000  0.19364839
#> WCS.HDY    -0.0404645157 -0.043113108  0.21147710  0.193648393  1.00000000
#> CL.EDM      0.0186420418  0.023041726 -0.02038525  0.001167407 -0.00711639
#> MSW.EDM     0.0003871251  0.001000149  0.30785859  0.270345986  0.30788781
#> Mars.CLO01 -0.0130463483 -0.006241277  0.33189503  0.358356635  0.17250132
#> Mars.CLO02 -0.0168388956 -0.013166126  0.37031431  0.405077418  0.19854543
#> LLS.SJA    -0.0129077174 -0.007769587  0.31407411  0.338674143  0.16396361
#> SYN.EDM     0.0163194376  0.014823299  0.29295759  0.272100182  0.31784004
#> C5.EDM      0.0034498064  0.001065234  0.27774379  0.264886858  0.25541288
#> DUB.EFS    -0.0470423924 -0.040598507  0.28931914  0.243616381  0.20924246
#> WTI.MEH    -0.0131203062 -0.008921146  0.34917433  0.367226535  0.18132500
#> WTI.MID     0.0175504004  0.019739576  0.28319810  0.296406822  0.16909446
#>                  CL.EDM       MSW.EDM   Mars.CLO01   Mars.CLO02      LLS.SJA
#> WTI.CMA01   0.018642042  0.0003871251 -0.013046348 -0.016838896 -0.012907717
#> WTI.CMA02   0.023041726  0.0010001489 -0.006241277 -0.013166126 -0.007769587
#> WCS.CUS    -0.020385247  0.3078585947  0.331895027  0.370314305  0.314074109
#> WCS.HOU     0.001167407  0.2703459863  0.358356635  0.405077418  0.338674143
#> WCS.HDY    -0.007116390  0.3078878090  0.172501318  0.198545430  0.163963610
#> CL.EDM      1.000000000 -0.0037967619 -0.011689953  0.005788827  0.005481465
#> MSW.EDM    -0.003796762  1.0000000000  0.266342922  0.327172086  0.302795061
#> Mars.CLO01 -0.011689953  0.2663429217  1.000000000  0.658280294  0.550797130
#> Mars.CLO02  0.005788827  0.3271720859  0.658280294  1.000000000  0.545199251
#> LLS.SJA     0.005481465  0.3027950606  0.550797130  0.545199251  1.000000000
#> SYN.EDM     0.016145102  0.7166179181  0.249799450  0.302983855  0.269042552
#> C5.EDM      0.002097182  0.4571170442  0.249586313  0.306855433  0.301966600
#> DUB.EFS    -0.019491678  0.3135125979  0.265228796  0.337759759  0.271405221
#> WTI.MEH     0.002993289  0.3509784492  0.518409769  0.564632370  0.579342134
#> WTI.MID    -0.012683684  0.2852215426  0.345391866  0.376365347  0.370698131
#>               SYN.EDM      C5.EDM     DUB.EFS      WTI.MEH     WTI.MID
#> WTI.CMA01  0.01631944 0.003449806 -0.04704239 -0.013120306  0.01755040
#> WTI.CMA02  0.01482330 0.001065234 -0.04059851 -0.008921146  0.01973958
#> WCS.CUS    0.29295759 0.277743788  0.28931914  0.349174330  0.28319810
#> WCS.HOU    0.27210018 0.264886858  0.24361638  0.367226535  0.29640682
#> WCS.HDY    0.31784004 0.255412877  0.20924246  0.181325000  0.16909446
#> CL.EDM     0.01614510 0.002097182 -0.01949168  0.002993289 -0.01268368
#> MSW.EDM    0.71661792 0.457117044  0.31351260  0.350978449  0.28522154
#> Mars.CLO01 0.24979945 0.249586313  0.26522880  0.518409769  0.34539187
#> Mars.CLO02 0.30298385 0.306855433  0.33775976  0.564632370  0.37636535
#> LLS.SJA    0.26904255 0.301966600  0.27140522  0.579342134  0.37069813
#> SYN.EDM    1.00000000 0.429123345  0.28179737  0.325000178  0.26919860
#> C5.EDM     0.42912335 1.000000000  0.30838757  0.334823700  0.26891822
#> DUB.EFS    0.28179737 0.308387568  1.00000000  0.310955931  0.27332748
#> WTI.MEH    0.32500018 0.334823700  0.31095593  1.000000000  0.50581953
#> WTI.MID    0.26919860 0.268918217  0.27332748  0.505819527  1.00000000
#> 
#> $coVaR
#> # A tibble: 15 × 16
#>    series     WTI.CMA01 WTI.CMA02   WCS.CUS   WCS.HOU  WCS.HDY   CL.EDM  MSW.EDM
#>    <chr>          <dbl>     <dbl>     <dbl>     <dbl>    <dbl>    <dbl>    <dbl>
#>  1 WTI.CMA01   2.40       2.21    -0.00819   -5.82e-3 -6.78e-2  2.08e-3  5.35e-4
#>  2 WTI.CMA02   2.21       2.21    -0.0117    -8.94e-3 -6.93e-2  2.46e-3  1.33e-3
#>  3 WCS.CUS    -0.00819   -0.0117   0.215      1.41e-1  1.06e-1 -6.80e-4  1.27e-1
#>  4 WCS.HOU    -0.00582   -0.00894  0.141      2.29e-1  1.00e-1  4.02e-5  1.15e-1
#>  5 WCS.HDY    -0.0678    -0.0693   0.106      1.00e-1  1.17e+0 -5.53e-4  2.97e-1
#>  6 CL.EDM      0.00208    0.00246 -0.000680   4.02e-5 -5.53e-4  5.16e-3 -2.43e-4
#>  7 MSW.EDM     0.000535   0.00133  0.127      1.15e-1  2.97e-1 -2.43e-4  7.95e-1
#>  8 Mars.CLO01 -0.0138    -0.00635  0.105      1.17e-1  1.28e-1 -5.75e-4  1.62e-1
#>  9 Mars.CLO02 -0.00989   -0.00742  0.0651     7.35e-2  8.13e-2  1.58e-4  1.11e-1
#> 10 LLS.SJA    -0.0356    -0.0206   0.259      2.89e-1  3.15e-1  7.01e-4  4.80e-1
#> 11 SYN.EDM     0.0232     0.0202   0.125      1.19e-1  3.15e-1  1.06e-3  5.85e-1
#> 12 C5.EDM      0.00418    0.00124  0.101      9.92e-2  2.16e-1  1.18e-4  3.19e-1
#> 13 DUB.EFS    -0.0321    -0.0266   0.0591     5.13e-2  9.95e-2 -6.16e-4  1.23e-1
#> 14 WTI.MEH    -0.0136    -0.00888  0.108      1.18e-1  1.31e-1  1.44e-4  2.09e-1
#> 15 WTI.MID     0.0192     0.0207   0.0928     1.00e-1  1.29e-1 -6.44e-4  1.80e-1
#> # … with 8 more variables: Mars.CLO01 <dbl>, Mars.CLO02 <dbl>, LLS.SJA <dbl>,
#> #   SYN.EDM <dbl>, C5.EDM <dbl>, DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>
#> 
#> $sims
#> # A tibble: 10 × 15
#>    WTI.CMA01 WTI.CMA02 WCS.CUS  WCS.HOU WCS.HDY   CL.EDM MSW.EDM Mars.CLO01
#>        <dbl>     <dbl>   <dbl>    <dbl>   <dbl>    <dbl>   <dbl>      <dbl>
#>  1    0.881     0.659  -0.681   0.487    0.679  -0.0230   -0.147      0.520
#>  2    1.78      1.85    0.589  -0.00119  1.05   -0.0512    0.846     -0.170
#>  3   -1.63     -2.01    0.0446 -0.365    0.453  -0.00705  -0.475     -0.654
#>  4   -1.42     -1.69    0.0761  0.458    1.26    0.0756    0.487      0.891
#>  5    1.35      0.702  -0.612  -0.319    0.0966 -0.0452   -0.364     -0.955
#>  6   -0.540    -0.358   0.392   0.656    0.881   0.0419    1.40       1.21 
#>  7    0.136    -0.230   0.279   0.742    0.702  -0.0671    1.04       0.704
#>  8   -1.25     -1.39    0.211  -0.395    0.458   0.0480   -0.709      0.189
#>  9   -0.198    -0.0670  0.185   0.416   -1.60    0.0920   -0.843      0.126
#> 10   -0.0400   -0.0272 -0.542  -0.203   -0.693  -0.0126    0.145      0.123
#> # … with 7 more variables: Mars.CLO02 <dbl>, LLS.SJA <dbl>, SYN.EDM <dbl>,
#> #   C5.EDM <dbl>, DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>
#>