Multivariate Normal from historical dataset

Generates multivariate random epsilons using absolute returns.

simMultivariates(nsims = 10, x, s0 = NULL)

Arguments

nsims

Number of simulations. Defaults to 10

x

Wide data frame of prices with date as first column.

s0

Vector of starting value for each variables. Defaults to NULL with zero.

Value

List of means, sds, covariance matrix, correlation matrix and simulated values

Author

Philippe Cote

Examples

simMultivariates(nsims = 10, x = RTL::fizdiffs, s0 = NULL)
#> $ave
#> # A tibble: 1 × 15
#>   WTI.CMA01 WTI.CM…¹  WCS.CUS  WCS.HOU  WCS.HDY   CL.EDM MSW.EDM Mars.…² Mars.…³
#>       <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>   <dbl>
#> 1    0.0165   0.0153 -0.00281 -0.00258 -0.00417 -4.41e-5 0.00245 0.00148 0.00109
#> # … with 6 more variables: LLS.SJA <dbl>, SYN.EDM <dbl>, C5.EDM <dbl>,
#> #   DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>, and abbreviated variable names
#> #   ¹​WTI.CMA02, ²​Mars.CLO01, ³​Mars.CLO02
#> 
#> $sd
#> # A tibble: 1 × 15
#>   WTI.C…¹ WTI.C…² WCS.CUS WCS.HOU WCS.HDY CL.EDM MSW.EDM Mars.…³ Mars.…⁴ LLS.SJA
#>     <dbl>   <dbl>   <dbl>   <dbl>   <dbl>  <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1    1.64    1.57   0.487   0.504    1.05 0.0691   0.868   0.679   0.377    1.75
#> # … with 5 more variables: SYN.EDM <dbl>, C5.EDM <dbl>, DUB.EFS <dbl>,
#> #   WTI.MEH <dbl>, WTI.MID <dbl>, and abbreviated variable names ¹​WTI.CMA01,
#> #   ²​WTI.CMA02, ³​Mars.CLO01, ⁴​Mars.CLO02
#> 
#> $corMat
#>               WTI.CMA01     WTI.CMA02      WCS.CUS     WCS.HOU     WCS.HDY
#> WTI.CMA01   1.000000000  0.9600417675 -0.026791907 -0.01688374 -0.04529817
#> WTI.CMA02   0.960041768  1.0000000000 -0.031957737 -0.02173117 -0.04767292
#> WCS.CUS    -0.026791907 -0.0319577365  1.000000000  0.66061594  0.19861599
#> WCS.HOU    -0.016883744 -0.0217311712  0.660615939  1.00000000  0.18902762
#> WCS.HDY    -0.045298166 -0.0476729236  0.198615988  0.18902762  1.00000000
#> CL.EDM     -0.004353840 -0.0041758676  0.001431385  0.01436687  0.01945278
#> MSW.EDM     0.006649877  0.0078779168  0.296931160  0.26654419  0.28889924
#> Mars.CLO01 -0.008325333 -0.0031842036  0.315809588  0.33712764  0.14729814
#> Mars.CLO02 -0.010799378 -0.0085067053  0.355756981  0.38570591  0.16565859
#> LLS.SJA    -0.008958644 -0.0043036571  0.297585295  0.31802776  0.14573375
#> SYN.EDM     0.025361209  0.0256185287  0.278713203  0.26939350  0.29022072
#> C5.EDM      0.009938843  0.0080804793  0.271253817  0.26639117  0.22998952
#> DUB.EFS    -0.054815936 -0.0480195727  0.295058868  0.26037586  0.17682274
#> WTI.MEH    -0.004408685 -0.0003411529  0.331883278  0.35656218  0.14566535
#> WTI.MID     0.019938954  0.0212883365  0.278590652  0.29305099  0.13493222
#>                  CL.EDM      MSW.EDM   Mars.CLO01   Mars.CLO02      LLS.SJA
#> WTI.CMA01  -0.004353840  0.006649877 -0.008325333 -0.010799378 -0.008958644
#> WTI.CMA02  -0.004175868  0.007877917 -0.003184204 -0.008506705 -0.004303657
#> WCS.CUS     0.001431385  0.296931160  0.315809588  0.355756981  0.297585295
#> WCS.HOU     0.014366871  0.266544188  0.337127643  0.385705908  0.318027764
#> WCS.HDY     0.019452782  0.288899241  0.147298143  0.165658589  0.145733749
#> CL.EDM      1.000000000 -0.018487095 -0.022529530 -0.006758497  0.008856564
#> MSW.EDM    -0.018487095  1.000000000  0.244061281  0.311443782  0.283491401
#> Mars.CLO01 -0.022529530  0.244061281  1.000000000  0.642185797  0.529354542
#> Mars.CLO02 -0.006758497  0.311443782  0.642185797  1.000000000  0.527385298
#> LLS.SJA     0.008856564  0.283491401  0.529354542  0.527385298  1.000000000
#> SYN.EDM    -0.004173107  0.729624450  0.237267240  0.298203233  0.262015182
#> C5.EDM     -0.012660044  0.459482463  0.232526521  0.291882147  0.290017181
#> DUB.EFS     0.003687964  0.305866612  0.245644634  0.319318100  0.261975886
#> WTI.MEH    -0.011195411  0.323198898  0.498337713  0.539365586  0.578540221
#> WTI.MID    -0.006423998  0.277300415  0.340708490  0.368307889  0.381689234
#>                 SYN.EDM       C5.EDM      DUB.EFS       WTI.MEH      WTI.MID
#> WTI.CMA01   0.025361209  0.009938843 -0.054815936 -0.0044086851  0.019938954
#> WTI.CMA02   0.025618529  0.008080479 -0.048019573 -0.0003411529  0.021288337
#> WCS.CUS     0.278713203  0.271253817  0.295058868  0.3318832777  0.278590652
#> WCS.HOU     0.269393500  0.266391173  0.260375863  0.3565621827  0.293050986
#> WCS.HDY     0.290220723  0.229989519  0.176822743  0.1456653472  0.134932221
#> CL.EDM     -0.004173107 -0.012660044  0.003687964 -0.0111954109 -0.006423998
#> MSW.EDM     0.729624450  0.459482463  0.305866612  0.3231988978  0.277300415
#> Mars.CLO01  0.237267240  0.232526521  0.245644634  0.4983377129  0.340708490
#> Mars.CLO02  0.298203233  0.291882147  0.319318100  0.5393655860  0.368307889
#> LLS.SJA     0.262015182  0.290017181  0.261975886  0.5785402211  0.381689234
#> SYN.EDM     1.000000000  0.438887111  0.268855123  0.3097122186  0.261076593
#> C5.EDM      0.438887111  1.000000000  0.304282580  0.3219650861  0.263675892
#> DUB.EFS     0.268855123  0.304282580  1.000000000  0.2879865763  0.272165115
#> WTI.MEH     0.309712219  0.321965086  0.287986576  1.0000000000  0.514409242
#> WTI.MID     0.261076593  0.263675892  0.272165115  0.5144092423  1.000000000
#> 
#> $coVaR
#> # A tibble: 15 × 16
#>    series     WTI.CMA01 WTI.CMA02    WCS.CUS  WCS.HOU  WCS.HDY   CL.EDM  MSW.EDM
#>    <chr>          <dbl>     <dbl>      <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#>  1 WTI.CMA01   2.68      2.47     -0.0213    -1.39e-2 -0.0781  -4.92e-4  0.00944
#>  2 WTI.CMA02   2.47      2.47     -0.0244    -1.72e-2 -0.0789  -4.53e-4  0.0107 
#>  3 WCS.CUS    -0.0213   -0.0244    0.237      1.62e-1  0.102    4.82e-5  0.125  
#>  4 WCS.HOU    -0.0139   -0.0172    0.162      2.54e-1  0.100    5.00e-4  0.117  
#>  5 WCS.HDY    -0.0781   -0.0789    0.102      1.00e-1  1.11     1.42e-3  0.264  
#>  6 CL.EDM     -0.000492 -0.000453  0.0000482  5.00e-4  0.00142  4.78e-3 -0.00111
#>  7 MSW.EDM     0.00944   0.0107    0.125      1.17e-1  0.264   -1.11e-3  0.754  
#>  8 Mars.CLO01 -0.00925  -0.00340   0.104      1.15e-1  0.105   -1.06e-3  0.144  
#>  9 Mars.CLO02 -0.00666  -0.00504   0.0653     7.32e-2  0.0658  -1.76e-4  0.102  
#> 10 LLS.SJA    -0.0257   -0.0118    0.254      2.81e-1  0.269    1.07e-3  0.431  
#> 11 SYN.EDM     0.0383    0.0371    0.125      1.25e-1  0.282   -2.66e-4  0.585  
#> 12 C5.EDM      0.0128    0.00996   0.104      1.05e-1  0.190   -6.87e-4  0.313  
#> 13 DUB.EFS    -0.0418   -0.0351    0.0669     6.11e-2  0.0868   1.19e-4  0.124  
#> 14 WTI.MEH    -0.00478  -0.000355  0.107      1.19e-1  0.102   -5.13e-4  0.186  
#> 15 WTI.MID     0.0228    0.0234    0.0947     1.03e-1  0.0993  -3.10e-4  0.168  
#> # … with 8 more variables: Mars.CLO01 <dbl>, Mars.CLO02 <dbl>, LLS.SJA <dbl>,
#> #   SYN.EDM <dbl>, C5.EDM <dbl>, DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>
#> 
#> $sims
#> # A tibble: 10 × 15
#>    WTI.CMA01 WTI.CM…¹  WCS.CUS  WCS.HOU WCS.HDY   CL.EDM MSW.EDM Mars.…² Mars.…³
#>        <dbl>    <dbl>    <dbl>    <dbl>   <dbl>    <dbl>   <dbl>   <dbl>   <dbl>
#>  1    -1.65    -3.10  -0.114    0.117   -0.586  -0.118    -1.38  -0.334  -0.0334
#>  2     1.70     1.06   0.304   -0.00885  0.0450 -0.0119   -0.396 -1.43   -0.246 
#>  3     2.09     1.44   0.102   -0.235    1.05    0.0721   -0.125  0.669   0.310 
#>  4    -2.44    -2.47  -0.559   -0.673    0.332  -0.106     0.704 -0.974   0.0241
#>  5    -2.77    -2.54  -0.671   -0.430   -1.11   -0.0606   -0.616  0.0116  0.272 
#>  6    -1.43    -1.49   0.329    0.181    1.27   -0.136    -0.134 -0.473   0.508 
#>  7     0.963    0.852 -0.265   -0.322   -0.920   0.00888  -0.207  0.0249 -0.286 
#>  8     0.837    0.576 -0.153    0.384   -0.336   0.0133    1.91   0.968   0.967 
#>  9    -0.763   -1.65  -0.00287  0.381   -0.193   0.0254    0.865  0.864   0.442 
#> 10     2.18     2.76  -0.511   -0.708   -0.461   0.0180    0.288 -0.126  -0.281 
#> # … with 6 more variables: LLS.SJA <dbl>, SYN.EDM <dbl>, C5.EDM <dbl>,
#> #   DUB.EFS <dbl>, WTI.MEH <dbl>, WTI.MID <dbl>, and abbreviated variable names
#> #   ¹​WTI.CMA02, ²​Mars.CLO01, ³​Mars.CLO02
#>